persister/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
kafka-python==2.0.1
djangorestframework==3.12.1
requests==2.24.0
